Move node v1 api to data_set_manager.
